requirements.txt
absl-py==0.13.0
boto3==1.18.44
botocore==1.21.44
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cycler==0.10.0
decorator==5.1.0
Flask==2.0.1
Flask-Cors==3.0.10
future==0.18.2
geocoder==1.38.1
google-auth==1.35.0
google-auth-oauthlib==0.4.6
greenlet==1.1.1
grpcio==1.40.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.3.56
pandas==1.3.3
Pillow==8.3.2
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
ratelim==0.1.6
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
SQLAlchemy==1.4.23
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
torch==1.9.1
torchaudio==0.9.1
torchvision==0.10.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1